Changes to ensure clean siteId column drop

+++ src/migrations/m220104_164702_simplify_cache_strategy.php
@@ -2,6 +2,7 @@
 
 namespace internetztube\elementRelations\migrations;
 
+use Craft;
 use craft\db\Migration;
 use internetztube\elementRelations\records\ElementRelationsRecord;
 
@@ -10,21 +11,38 @@ class m220104_164702_simplify_cache_strategy extends Migration
     public function safeUp()
     {
         $table = ElementRelationsRecord::tableName();
+        // drop previous foreign key on 'siteId` which was built by the install so we can drop the column
+        foreach (Craft::$app->db->getSchema()->getTableForeignKeys($table) as $foreignKey) {
+            if (in_array('siteId',$foreignKey->columnNames)) {
+                $keyName = $foreignKey->name;
+                $this->dropForeignKey($foreignKey->name, $table);
+            }
+        }
+
         // drop previous index on 'siteId` which was built by the install so we can drop the column
         foreach (Craft::$app->db->getSchema()->findIndexes($table) as $name => $index) {
             if ($index['columns'] === ['siteId']) {
                 $this->dropIndex($name, $table);
             }
         }
-        $this->dropColumn($table, 'siteId');
-        $this->dropColumn($table, 'markup');
-        $this->createIndex(null, $table, ['relations', 'dateUpdated']);
+        if (Craft::$app->db->columnExists($table, 'siteId')) {
+            $this->dropColumn($table, 'siteId');
+        }
+        if (Craft::$app->db->columnExists($table, 'markup')) {
+            $this->dropColumn($table, 'markup');
+        }
+        // Adding an index on a text field in mysql requires a length
+        if (Craft::$app->db->getIsMysql()) {
+            $this->createIndex(null, $table, ['relations(250)', 'dateUpdated']);
+        } else {
+            $this->createIndex(null, $table, ['relations', 'dateUpdated']);
+        }
         return true;
     }
 
     public function safeDown()
     {
-        echo "m220104_132645_simplify_cache_strategy cannot be reverted.\n";
+        echo "m220104_164702_simplify_cache_strategy cannot be reverted.\n";
         return false;
     }
 }
